requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-mathfilters==1.0.0
executing==1.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==22.3.5
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==3.5.3
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
